+++ NEWS.md
@@ -5,7 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [Unreleased]
+
+
+## [1.1.0] - 2025-01-09
 
 ### Added
 
@@ -34,5 +38,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New CHANGELOG (`NEWS.md`)
 * Documentation for `logic_reducer()`, `no_milestones()`, `vgrepl()` and `simple_sort`
 
-[Unreleased]: https://github.com/TanguyBarthelemy/TBox/compare/v1.0.0...HEAD
-[1.0.0]: https://github.com/TanguyBarthelemy/TBox/releases/tag/v1.0.0
+[Unreleased]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/TanguyBarthelemy/IssueTrackeR/compare/v1.0.0...v1.1.0
+[1.0.0]: https://github.com/TanguyBarthelemy/IssueTrackeR/releases/tag/v1.0.0